+++ src/otp-field.ts
@@ -1,0 +1,1 @@
+export * from "./components/otp-field/index.js";